move enums to the corresponding spikewall/Object classes
