--- Enums/AdvertEventType.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum AdvertEventType
-    {
-        Roulette,
-        Character,
-        Shop
-    }
-}
