--- Enums/BoostItemType.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum BoostItem
-    {
-        ScoreBonus,
-        AssistTrampoline,
-        SubCharacter
-    }
-}
