--- Enums/Campaign.cs
@@ -1,25 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum Campaign
-    {
-        BankedRingBonus,
-        DailyMissionBonus,
-        ChaoRouletteCost,
-        GameItemCost,
-        CharacterUpgradeCost,
-        PurchaseAddRings,
-        JackpotValueBonus,
-        MileagePassingRingBonus,
-        PurchaseAddEnergies,
-        PurchaseAddRedRings,
-        PurchaseAddRedRingsNoChargeUser,
-        SendAddEnergies,
-        InviteCount,
-        PremiumRouletteOdds,
-        FreeWheelSpinCount,
-        ContinueCost,
-        PurchaseAddRaidEnergies
-    }
-}
