--- Enums/ChaoIDs.cs
@@ -1,102 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum ChaoID
-    {
-        HeroChao,
-        GoldChao,
-        DarkChao,
-        JewelChao,
-        NormalChao,
-        Omochao,
-        RCSpring,
-        BabyCyanWisp,
-        BabyIndigoWisp,
-        BabyYellowWisp,
-        RCPinwheel,
-        RCPiggyBank,
-        RCBalloon,
-        EggChao,
-        PumpkinChao,
-        SkullChao,
-        Yacker,
-        RCGoldenPiggyBank,
-        WizardChao,
-        HeroChaos,
-        Chip,
-        Caliburn,
-        RCMonkey,
-        RCMagnet,
-        RCUFO,
-        RCBattleCrusier,
-        EasterChao,
-        EasterBunny,
-        Merlina,
-        Shahra,
-        MagLv1,
-        Rappy,
-        Kuna,
-        PurplePapurisu,
-        Suketoudara,
-        Carbuncle,
-        Cartridge,
-        Genesis,
-        DeathEgg,
-        DesertChao,
-        MagicLamp,
-        ErazorDjinn,
-        RCGoldenAngel,
-        RCTornado,
-        Nightopian,
-        NiGHTS,
-        Reala,
-        BlowfishTransporter,
-        MarineChao,
-        MotherWisp,
-        Orca,
-        Chaos,
-        RCTornado2,
-        GreenCrystalMonsterS,
-        GreenCrystalMonsterL,
-        LightChaos,
-        RCTurtle,
-        DarkChaos,
-        RCPirateSpaceship,
-        RCSatellite,
-        StarShapedMissiles,
-        RCMoonMech,
-        GoldenGoose,
-        RedCrystalMonsterS,
-        RedCrystalMonsterL,
-        Boo,
-        HalloweenChao,
-        KingBoomBoo,
-        SonicOmochao,
-        TailsOmochao,
-        KnucklesOmochao,
-        RCAirship,
-        KingAuthursGhost,
-        HeavyBomb,
-        OPapa,
-        OpaOpa,
-        DarkQueen,
-        ChaoWalker,
-        Ideya,
-        ChristmasNightopian,
-        ChristmasNiGHTS,
-        SnowChao,
-        Yeti,
-        ChristmasYeti,
-        RCBomber,
-        RCFighter,
-        RCHelicopter,
-        RCHovercraft,
-        Orbot,
-        Cubot,
-        DarkChaoWalker,
-        HunkofMeat,
-        BlockBomb,
-        RCBlockFace
-    }
-}
