--- Enums/ChaoRarity.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum ChaoRarity
-    {
-        Normal,
-        Rare,
-        SRare,
-        None
-    }
-}
