--- Enums/ChaoWheelType.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum ChaoWheelType
-    {
-        Normal,
-        Special
-    }
-}
