--- Enums/CharaType.cs
@@ -1,38 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum CharaType
-    {
-        Sonic,
-        Tails,
-        Knuckles,
-        Amy,
-        Shadow,
-        Blaze,
-        Rouge,
-        Omega,
-        Big,
-        Cream,
-        Espio,
-        Charmy,
-        Vector,
-        Silver,
-        MetalSonic,
-        AmitieAmy,
-        ClassicSonic,
-        Tikal,
-        GothicAmy,
-        HalloweenShadow,
-        HalloweenRouge,
-        HalloweenOmega,
-        Mephiles,
-        PSISilver,
-        ChristmasSonic,
-        ChristmasTails,
-        ChristmasKnuckles,
-        Werehog,
-        Sticks
-    }
-}
-
