--- Enums/CollectEventType.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum CollectEventType
-    {
-        GetAnimals,
-        GetRing,
-        RunDistance
-    }
-}
