--- Enums/EventIDs.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum EventID
-    {
-        SpecialStage,
-        RaidBoss,
-        CollectObject,
-        Gacha,
-        Advert,
-        Quick,
-        BGM
-    }
-}
