--- Enums/ItemIDs.cs
@@ -1,68 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum ItemID
-    {
-        BoostScore,
-        BoostTrampoline,
-        BoostSubChara,
-        Invincible,
-        Barrier,
-        Magnet,
-        Trampoline,
-        Combo,
-        Laser,
-        Drill,
-        Asteroid,
-        RingBonus,
-        DistanceBonus,
-        AnimalBonus,
-        PackedInvincible0,
-        PackedBarrier0,
-        PackedMagnet0,
-        PackedTrampoline0,
-        PackedCombo0,
-        PackedLaser0,
-        PackedDrill0,
-        PackedAsteroid0,
-        PackedRingBonus0,
-        PackedDistanceBonus0,
-        PackedAnimalBonus0,
-        PackedInvincible1,
-        PackedBarrier1,
-        PackedMagnet1,
-        PackedTrampoline1,
-        PackedCombo1,
-        PackedLaser1,
-        PackedDrill1,
-        PackedAsteroid1,
-        PackedRingBonus1,
-        PackedDistanceBonus1,
-        PackedAnimalBonus1,
-        RedRing,
-        RedRing0,
-        RedRing1,
-        RedRing2,
-        RedRing3,
-        RedRing4,
-        Ring,
-        Ring0,
-        Ring1,
-        Ring2,
-        Ring3,
-        Ring4,
-        Ring5,
-        Energy,
-        Energy0,
-        Energy1,
-        Energy2,
-        Energy3,
-        Energy4,
-        RaidbossEnergy,
-        RaidbossRing,
-        ItemRouletteTicket,
-        PremiumRouletteTicket
-    }
-
-}
