--- Enums/RankingLeague.cs
@@ -1,29 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum RankingLeague
-    {
-        F_M,
-        F, 
-        F_P,
-        E_M,
-        E,
-        E_P,
-        D_M,
-        D,
-        D_P,
-        C_M,
-        C,
-        C_P,
-        B_M,
-        B,
-        B_P,
-        A_M,
-        A,
-        A_P,
-        S_M,
-        S,
-        S_P
-    }
-}
