--- Enums/RankingMode.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum RankingMode
-    {
-        Endless,
-        Quick
-    }
-}
