--- Enums/UpgradeAbility.cs
@@ -1,18 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum UpgradeAbility
-    {
-        Invincible,
-        Magnet,
-        Trampoline,
-        Combo,
-        Laser,
-        Drill,
-        Asteroid,
-        RingBonus,
-        DistanceBonus,
-        AnimalBonus
-    }
-}
