--- Enums/WheelRank.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Enums
-{
-    public enum WheelRank
-    {
-        Normal,
-        Big,
-        Super
-    }
-}
